Renaming Swap to Transpose
After my first time using this extension, I noticed that "swap" is not
intuitive enough, I think "transpose" is the appropriate term for this
operation, and it is also the term used by spreadsheet software. The
README.md itself says “Swap (transpose) copy.”, which is another signal
that "swap" wasn't the best word choice.

In this commit, I just blindly rename Swap to Transpose. I haven't
tested the code in the browser, though. (But I don't expect any issue.)